third_party/blink/web_tests/external/wpt/editing/run/delete_2001-3000-expected.txt

This is a testharness.js-based test.
Found 24 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["delete",""]] "foo<blockquote>[]bar</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foobar"
[FAIL] [["delete",""]] "foo<blockquote><blockquote>[]bar</blockquote></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blockquote>bar</blockquote>" but got "foobar"
[FAIL] [["delete",""]] "foo<blockquote><div>[]bar</div></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div>bar</div>" but got "foobar"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<blockquote style=\\"color: blue\\">[]bar</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"color:rgb(0, 0, 255)\\">bar</div>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">bar</span>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<blockquote style=\\"color: blue\\">[]bar</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"color:rgb(0, 0, 255)\\">bar</div>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">bar</span>"
[FAIL] [["delete",""]] "foo<blockquote><blockquote><p>[]bar<p>baz</blockquote></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blockquote><p>bar</p><blockquote><p>baz</p></blockquote></blockquote>" but got "foobar<blockquote><blockquote><p>baz</p></blockquote></blockquote>"
[FAIL] [["delete",""]] "foo<blockquote><div><p>[]bar<p>baz</div></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div><p>bar</p><blockquote><p>baz</p></blockquote></div>" but got "foobar<blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<blockquote style=\\"color: blue\\"><p>[]bar<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"color:rgb(0, 0, 255)\\"><p>bar</p><blockquote><p>baz</p></blockquote></div>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">bar</span><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<blockquote style=\\"color: blue\\"><p>[]bar<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"color:rgb(0, 0, 255)\\"><p>bar</p><blockquote><p>baz</p></blockquote></div>" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">bar</span><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>"
[FAIL] [["delete",""]] "foo<blockquote><p><b>[]bar</b><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p><b>bar</b></p><blockquote><p>baz</p></blockquote>" but got "foo<b>bar</b><blockquote><p>baz</p></blockquote>"
[FAIL] [["delete",""]] "foo<blockquote><p><strong>[]bar</strong><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p><strong>bar</strong></p><blockquote><p>baz</p></blockquote>" but got "foo<strong>bar</strong><blockquote><p>baz</p></blockquote>"
[FAIL] [["delete",""]] "foo<blockquote><p><span>[]bar</span><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p><span>bar</span></p><blockquote><p>baz</p></blockquote>" but got "foobar<blockquote><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<blockquote><ol><li>[]bar</ol></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blockquote><div>bar</div></blockquote><p>extra</p>" but got "foobar<p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<blockquote><ol><li>[]bar</ol></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blockquote><p>bar</p></blockquote><p>extra</p>" but got "foobar<p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "foo<blockquote>bar<ol><li>[]baz</ol>quz</blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blockquote>bar<div>baz</div>quz</blockquote><p>extra</p>" but got "foo<blockquote>barbaz<br>quz</blockquote><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "foo<blockquote>bar<ol><li>[]baz</ol>quz</blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blockquote>bar<p>baz</p>quz</blockquote><p>extra</p>" but got "foo<blockquote>barbaz<br>quz</blockquote><p>extra</p>"
[FAIL] [["delete",""]] "foo<blockquote><ol><li>bar</li><ol><li>[]baz</ol><li>quz</ol></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<blockquote><ol><li>bar</li><li>baz</li><li>quz</li></ol></blockquote><p>extra</p>" but got "foo<blockquote><ol><li>barbaz</li><li>quz</li></ol></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p style=color:blue>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"color:rgb(0, 0, 255)\\">foo</span>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p style=color:blue>foo<p>[]bar" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p style=color:blue>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"#0000ff\\">foo</font>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["delete",""]] "<p style=color:blue>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"color:rgb(0, 0, 255)\\">foo</span>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p style=color:blue>foo<p>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"#0000ff\\">foo</font>bar</p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foobar</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p style=color:blue>foo<p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#a52a2a\\">bar</font></p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p style=color:blue>foo<p style=color:brown>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"color:rgb(0, 0, 255)\\">foo<font color=\\"#a52a2a\\">bar</font></p>" but got "<p style=\\"color:rgb(0, 0, 255)\\">foo<span style=\\"color:rgb(165, 42, 42)\\">bar</span></p>"
Harness: the test ran to completion.
